implement jwt token provider

+++ src/features/auth/infrastructure/entities/tokens_pair.rs
@@ -1,6 +1,6 @@
 pub struct TokensPair {
     pub access_token: String,
     pub refresh_token: String,
-    pub access_token_exp: usize,
-    pub refresh_token_exp: usize,
+    pub access_token_exp: u64,
+    pub refresh_token_exp: u64,
 }
